+++ src/hyp3_autorift/image.py
@@ -42,6 +42,8 @@ COLOR_MAP = np.array(
 def make_browse(
     out_file: Path, data: np.ndarray, min_value: Optional[float] = None, max_value: Optional[float] = 625.0
 ) -> Path:
+    data = np.squeeze(data)
+
     data_values = COLOR_MAP[:, 0]
     pchip = PchipInterpolator(data_values, np.linspace(0, 1, len(data_values)))
     image = pchip(np.clip(data, min_value, max_value))